operation_fcts: Cql.g: add grammar for Add, Subtract and Multiply operations
When an Add or Subtract sign is identified, both sides of the operation
will be processed and sent to an operation_fct. The function will try to
identify the return type through type hints, cast operands apropriately
and return the right type for the operation. Multiply and Divisions take
precedence over Add and Subtract, thus they are processed first.

Signed-off-by: Juliana Oliveira <juliana@scylladb.com>